+++ thirdeye-detectionpipeline/src/main/java/ai/startree/thirdeye/detectionpipeline/operator/ForkJoinResultItem.java
@@ -14,26 +14,87 @@
 
 package ai.startree.thirdeye.detectionpipeline.operator;
 
+import ai.startree.thirdeye.detectionpipeline.DetectionPipelineContext;
 import ai.startree.thirdeye.spi.datalayer.dto.EnumerationItemDTO;
+import ai.startree.thirdeye.spi.datalayer.dto.MergedAnomalyResultDTO;
+import ai.startree.thirdeye.spi.detection.model.TimeSeries;
 import ai.startree.thirdeye.spi.detection.v2.OperatorResult;
+import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
+import java.util.stream.Collectors;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 public class ForkJoinResultItem {
 
   private final EnumerationItemDTO enumerationItem;
   private final Map<String, OperatorResult> results;
+  private final DetectionPipelineContext detectionPipelineContext;
 
   public ForkJoinResultItem(final EnumerationItemDTO enumerationItem,
-      final Map<String, OperatorResult> results) {
+      final Map<String, OperatorResult> results,
+      DetectionPipelineContext detectionPipelineContext) {
     this.enumerationItem = enumerationItem;
-    this.results = results;
-  }
-
-  public EnumerationItemDTO getEnumerationItem() {
-    return enumerationItem;
+    this.detectionPipelineContext = detectionPipelineContext;
+    this.results = process(results);
   }
 
   public Map<String, OperatorResult> getResults() {
     return results;
   }
+
+
+  private Map<String, OperatorResult> process(final Map<String, OperatorResult> outputs) {
+    if (outputs == null) {
+      return null;
+    }
+
+    return outputs.entrySet()
+        .stream()
+        .collect(Collectors.toMap(Entry::getKey, e -> conditionalClone(e.getValue()), (a, b) -> b));
+  }
+
+  /**
+   * The main intent of cloning the operator result is to skip the timeseries if not required.
+   * This intentionally loses the src result and forgets references so that they can be GC'd later.
+   *
+   * @param src operator result
+   * @return cloned operator result according to {@link DetectionPipelineContext} instance
+   */
+  private OperatorResult conditionalClone(final OperatorResult src) {
+    final long lastTimestamp = src.getLastTimestamp();
+    final List<MergedAnomalyResultDTO> anomalies = src.getAnomalies();
+    final TimeSeries timeSeries = detectionPipelineContext.isPreserveOutputDataFrames()
+        ? src.getTimeseries()
+        : null;
+    final Map<String, List> rawData = src.getRawData();
+
+    return new OperatorResult() {
+      @Override
+      public long getLastTimestamp() {
+        return lastTimestamp;
+      }
+
+      @Override
+      public @Nullable List<MergedAnomalyResultDTO> getAnomalies() {
+        return anomalies;
+      }
+
+      @Override
+      public @Nullable EnumerationItemDTO getEnumerationItem() {
+        return enumerationItem;
+      }
+
+      @Override
+      public @Nullable Map<String, List> getRawData() {
+        return rawData;
+      }
+
+      @Override
+      public @Nullable TimeSeries getTimeseries() {
+        return timeSeries;
+      }
+    };
+  }
+
 }