+++ thirdeye-server/src/main/java/ai/startree/thirdeye/core/AppAnalyticsService.java
@@ -13,8 +13,16 @@
  */
 package ai.startree.thirdeye.core;
 
+import static ai.startree.thirdeye.spi.detection.AnomalyFeedbackType.ANOMALY;
+import static ai.startree.thirdeye.spi.detection.AnomalyFeedbackType.ANOMALY_EXPECTED;
+import static ai.startree.thirdeye.spi.detection.AnomalyFeedbackType.ANOMALY_NEW_TREND;
+import static ai.startree.thirdeye.spi.detection.AnomalyFeedbackType.NOT_ANOMALY;
+import static ai.startree.thirdeye.spi.detection.AnomalyFeedbackType.NO_FEEDBACK;
+
 import ai.startree.thirdeye.alert.AlertTemplateRenderer;
 import ai.startree.thirdeye.spi.api.AnomalyStatsApi;
+import ai.startree.thirdeye.spi.datalayer.DaoFilter;
+import ai.startree.thirdeye.spi.datalayer.Predicate;
 import ai.startree.thirdeye.spi.datalayer.bao.AlertManager;
 import ai.startree.thirdeye.spi.datalayer.bao.MergedAnomalyResultManager;
 import ai.startree.thirdeye.spi.datalayer.dto.AlertDTO;
@@ -50,10 +58,8 @@ public class AppAnalyticsService {
   private final AlertTemplateRenderer renderer;
   private final MergedAnomalyResultManager anomalyManager;
 
-  public Supplier<ConfusionMatrix> confusionMatrixSupplier =
-      Suppliers.memoizeWithExpiration(this::computeConfusionMatrixForAnomalies, 5, TimeUnit.MINUTES)::get;
   public Supplier<List<AnomalyFeedback>> anomalyFeedbacksSupplier =
-      Suppliers.memoizeWithExpiration(this::getAnomalyFeedbacks, 5, TimeUnit.MINUTES)::get;
+      Suppliers.memoizeWithExpiration(this::getAllAnomalyFeedbacks, 5, TimeUnit.MINUTES)::get;
   public Supplier<Set<MonitoredMetricWrapper>> uniqueMonitoredMetricsSupplier =
       Suppliers.memoizeWithExpiration(this::getUniqueMonitoredMetrics, 5, TimeUnit.MINUTES)::get;
 
@@ -75,13 +81,13 @@ public class AppAnalyticsService {
     metricRegistry.register("anomalyPrecision", new CachedGauge<Double>(1, TimeUnit.HOURS) {
       @Override
       protected Double loadValue() {
-        return confusionMatrixSupplier.get().getPrecision();
+        return computeConfusionMatrixForAnomalies(null).getPrecision();
       }
     });
     metricRegistry.register("anomalyResponseRate", new CachedGauge<Double>(1, TimeUnit.HOURS) {
       @Override
       protected Double loadValue() {
-        return confusionMatrixSupplier.get().getResponseRate();
+        return computeConfusionMatrixForAnomalies(null).getResponseRate();
       }
     });
   }
@@ -115,51 +121,55 @@ public class AppAnalyticsService {
         .setDatasource(metadata.getDatasource().getName());
   }
 
-  public ConfusionMatrix computeConfusionMatrixForAnomalies() {
+  public ConfusionMatrix computeConfusionMatrixForAnomalies(DaoFilter filter) {
     final ConfusionMatrix matrix = new ConfusionMatrix();
-    matrix.addUnclassified((int) anomalyManager.countParentAnomaliesWithoutFeedback());
-    final List<AnomalyFeedback> feedbacks = anomalyFeedbacksSupplier.get();
-    for (final AnomalyFeedback feedback : feedbacks) {
-      final AnomalyFeedbackType type = feedback.getFeedbackType();
-      switch (type) {
-        case NO_FEEDBACK:
-          matrix.incUnclassified();
-          break;
-        case NOT_ANOMALY:
-          matrix.incFalsePositive();
-          break;
-        case ANOMALY:
-        case ANOMALY_EXPECTED:
-        case ANOMALY_NEW_TREND:
-          matrix.incTruePositive();
-          break;
-        default:
-          log.error("Unsupported feedback type: {}", type);
-      }
+    // filter to get anomalies without feedback
+    final Predicate predicate = Predicate.EQ("anomalyFeedbackId", 0L);
+    if (filter != null) {
+      filter.setPredicate(Predicate.AND(filter.getPredicate(), predicate));
+    } else {
+      filter = new DaoFilter().setPredicate(predicate);
     }
+    matrix.addUnclassified((int) anomalyManager.countParentAnomalies(filter));
+    final Map<AnomalyFeedbackType, Long> typeMap = aggregateFeedbackTypes(anomalyFeedbacksSupplier.get());
+    matrix.addUnclassified(Math.toIntExact(typeMap.get(NO_FEEDBACK)));
+    matrix.addFalsePositive(Math.toIntExact(typeMap.get(NOT_ANOMALY)));
+    matrix.addTruePositive(Math.toIntExact(typeMap.get(ANOMALY))
+        + Math.toIntExact(typeMap.get(ANOMALY_EXPECTED))
+        + Math.toIntExact(typeMap.get(ANOMALY_NEW_TREND)));
     return matrix;
   }
 
-  private List<AnomalyFeedback> getAnomalyFeedbacks() {
-    return anomalyManager.findParentAnomaliesWithFeedback().stream()
+  private List<AnomalyFeedback> getAllAnomalyFeedbacks() {
+    return getAnomalyFeedbacks(null);
+  }
+
+  private List<AnomalyFeedback> getAnomalyFeedbacks(final DaoFilter filter) {
+    return anomalyManager.findParentAnomaliesWithFeedback(filter).stream()
         .map(MergedAnomalyResultDTO::getFeedback)
         .collect(Collectors.toList());
   }
 
-  public AnomalyStatsApi computeAnomalyStats() {
-    List<AnomalyFeedback> feedbacks = anomalyFeedbacksSupplier.get();
-    Map<AnomalyFeedbackType, Long> feedbackStats = new HashMap<>();
-    for(AnomalyFeedbackType type : AnomalyFeedbackType.values()) {
+  public AnomalyStatsApi computeAnomalyStats(final DaoFilter filter) {
+    final List<AnomalyFeedback> allFeedbacks = filter == null
+        ? anomalyFeedbacksSupplier.get()
+        : getAnomalyFeedbacks(filter);
+    return new AnomalyStatsApi()
+        .setTotalCount(anomalyManager.countParentAnomalies(filter))
+        .setCountWithFeedback((long) allFeedbacks.size())
+        .setFeedbackStats(aggregateFeedbackTypes(allFeedbacks));
+  }
+
+  private Map<AnomalyFeedbackType, Long> aggregateFeedbackTypes(final List<AnomalyFeedback> feedbacks) {
+    final Map<AnomalyFeedbackType, Long> feedbackStats = new HashMap<>();
+    for(final AnomalyFeedbackType type : AnomalyFeedbackType.values()) {
       feedbackStats.put(type, 0L);
     }
     feedbacks.forEach(feedback -> {
-      AnomalyFeedbackType type = feedback.getFeedbackType();
-      long count = feedbackStats.get(type);
+      final AnomalyFeedbackType type = feedback.getFeedbackType();
+      final long count = feedbackStats.get(type);
       feedbackStats.put(type, count+1);
     });
-    return new AnomalyStatsApi()
-        .setTotalCount(anomalyManager.countParentAnomalies())
-        .setCountWithFeedback((long) feedbacks.size())
-        .setFeedbackStats(feedbackStats);
+    return feedbackStats;
   }
 }